refactor : 누락된 RacingCar 패키지에 포함
