+++ src/main/java/racingcar/application/service/RacingCarConverter.java
@@ -2,7 +2,7 @@ package racingcar.application.service;
 
 import java.util.Arrays;
 import java.util.List;
-import racingcar.RacingCar;
+import racingcar.racing.RacingCar;
 import racingcar.application.Converter;
 
 public class RacingCarConverter implements Converter<RacingCar> {
